factored out some common GPT2-like code to split_on_bpe_pairs #14
